# Performance: Larger BlockTile optimizations enable 1470+ TF FP8 on the "H800"-SXM